fix(e2e): address review feedback
- Split tests/flows/onboarding.spec.ts into api-onboarding.spec.ts and
  wc-onboarding.spec.ts so flows-api doesn't drag in the WalletConnect
  fixture (and WC_PROJECT_ID) just to run the API path. Drop the now
  unused `walletKind` option from both fixture files.
- Add an explicit `expect(...).toBeVisible({ timeout: 45_000 })` before
  clicking #tx-preview-confirm in the WC onboarding flow, matching the
  pattern used by the dedicated WC specs.
- Pass explicit longer timeouts to DashboardPage.runLitAction (60s) and
  getLitActionIpfsCid (30s) — the global expect timeout (15s) is too
  tight for Lit Action execution and CID hashing.
- Gate the WalletConnect display_uri DOM event re-emission in
  lit-static/wallet_connect.js on localhost so the pairing URI doesn't
  leak to other scripts loaded on a production origin.
- Update wallet-setup/basic.setup.ts comment to match reality (Synpress
  imports account #0 by default; the previous comment claimed #1).
- README: drop the stale flows-eoa references, document the new flows
  filename split, and mark Docker as optional now that Jaeger is.
